fix: `RangeError [ERR_OUT_OF_RANGE] exception when serializing transaction`
When `reservePaysFees` option is enabled, the transaction fee is deducted from the reserve's reward allocation when constructing the rewards payment transaction. This is a preference to ensure rewards due can be paid even if the delegate wallet is empty to start with.

Exception is raised due to negative transfer amount and the issue surfaces under a potentially rare condition:
1. reservePaysFees option enabled (default)
2. delegate reserve wallet is self voting,
3. in populated pay-order, reserve wallet voter allotment (payeeType=1) comes before reserve wallet reserve allotment (payeeType=0) in the array
4. reserve wallet voter allocation is less than the required transaction fee

The issue is now fixed and tx fee is deducted only from reserve's allocation (payeeType=0) and when it can cover the fee.
Furthermore, the algorithm is also improved in order to:
- Try subsequent reserve wallets if first one cannot cover the costs
- Log a warning if none of the reserve wallet allocations meet the criteria (in which case tx fee will be paid from the delegate wallet as if  `reservePaysFees=false`).